Revert "Rename isomorphism and layouts"
This reverts commit 22abf8505a740bb8c5f069e261cee8cc0ed8ab87.
