+++ src/isomorphism/mod.rs
@@ -11,6 +11,7 @@
 // under the License.
 
 #![allow(clippy::float_cmp)]
+#![allow(clippy::module_inception)]
 
 mod isomorphism;
