+++ src/lib.rs
@@ -20,9 +20,9 @@ mod digraph;
 mod dot_utils;
 mod generators;
 mod graph;
-mod isomorphism_algo;
+mod isomorphism;
 mod iterators;
-mod layout_algo;
+mod layout;
 mod matching;
 mod random_circuit;
 mod representation_algo;
@@ -37,8 +37,8 @@ use coloring::*;
 use connectivity::*;
 use core_number::*;
 use dag_algo::*;
-use isomorphism_algo::*;
-use layout_algo::*;
+use isomorphism::*;
+use layout::*;
 use matching::*;
 use random_circuit::*;
 use representation_algo::*;
